Added 'rvnc.nix'
Added 'rvnc.nix' to repository. This package downloads and installs the RealVNC VNC Viewer application directly from the RealVNC website. This package is intended for NixOS systems; other builds have not yet been tested. 

To install this package:

1. Download 'rvnc.nix'
2. Use a text editor like vim or nano to open the file
3. Double-check that all paths and directories are correct
4. Navigate to directory containing 'rvnc.nix' 
5. Run 'nix-build rvnc.nix'
6. Run './result/bin/rvnc'

NOTE: If you want to be able to launch rvnc from ANY directory, you'll need to add this package to your PATH. To do this:

1. Run '(sudo) nano ~/.bashrc'  (for Bash users) OR 'nano ~/.zshrc'  (for Zsh users) OR the appropriate, relevant command
2. Add the following line: 'export PATH=$PATH:/home/<yourusername>/<path/to/result/bin>'
3. You can now run 'rvnc' in your terminal from ANY location to launch the RealVNC VNC Viewer